Extract common global search aspects
